% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textych.R
\name{textych}
\alias{textych}
\title{Create a textych}
\usage{
textych(data, text = NULL, text_index = NULL, color = NULL, tooltip = NULL)
}
\arguments{
\item{data}{A table of data}

\item{text}{Input column of words}

\item{text_index}{Variable to group text parallels by}

\item{color}{Variable that assigns a color to word}

\item{tooltip}{Variable specifying tooltip}
}
\description{
Create a textych
}
