% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textych.R
\name{tt_style}
\alias{tt_style}
\title{Style a textych}
\usage{
tt_style(
  textych,
  text_size = 14,
  text_weight = "normal",
  text_font = "Arial",
  title_text_size = 14,
  title_text_weight = "bold",
  title_text_font = "Arial"
)
}
\arguments{
\item{textych}{An object of class 'textych'}

\item{text_size}{Size of text in parallels}

\item{text_weight}{Font-weight of text in parallels}

\item{text_font}{Font of text in parallels}

\item{title_text_size}{Size of title text}

\item{title_text_weight}{Font-weight of title text}

\item{title_text_font}{Font of title text}
}
\value{
An object of class 'textych'
}
\description{
Style a textych
}
